downloader: Set filename extensions correctly
The downloader assumes that all images will be JPEGs, setting the
filename extension accordingly. In reality, the resulting images may be
of many filetypes. Determine the type of image downloaded by checking
the magic number at the start of the image data, and then setting the
filename accordingly.

Currently, this only checks for some JPEG and PNG formats. if another
unknown format is encountered, the downloader will append the generic
extension ".img".