Begin implementation of download/filter/classify pipeline
Implement a first cut of the download/filter/classifier pipeline that
will enable automated iteration across different hyperparameter settings
for the classifier, as well as dataset collection/generation.

Currently, a single, non-filtered pass through the pipeline is
functional. The undesirable image filter still needs to be ported into
the repository, and logic to generate different permutations of
parameters has yet to be implemented.