classifier: fix type mismatch for accuracy comparison
The _train_one and _test_network functions return dictionaries of the
form

	{'loss': float, 'accuracy': float}

but the invocation of these functions assigns the return value to
multiple variables, as if it were a tuple. Consequently, the best vs
current accuracy comparison is acutally comparing the keys of the
dictionary. Fix this to correctly index into the returned dictionary.