classifier: deconflict evaluate() interface
Update the evaluate() API to return a tuple of train, validation, and
test dictionaries, each containing keys 'loss' and 'accuracy'.

Ultimately the return value here needs to be "actionable" , i.e.,
something from which a maximum can be selected to find the ideal set of
hyperparameters. Since a full set of loss/accuracy values is not
directly useful for this purpose, return the average loss and accuracy
for each set of values.