filter: support non-CUDA environments
The .pth file containing the serialized filter model apparently is tied
to the device used when the model was trained. When CUDA is not
available, loading the model will result in an exception. Check for
CUDA, and if it is not available, use torch.load()'s 'map_location'
keyword argument to use the CPU instead.